refactor: create grants service & renames
